Adding experimental function 'mapRow'
Added 'mapRow' (found in ResultSetMapper.kt) and its supporting enums, functions.

This 'mapRow' function reduces the boilerplate of manual mapping code required and allows for a quick turnaround for simpler use cases when result sets closely match their expected return Kotlin types. Please see the initial Kdoc on the function, as well as the unit tests (ResultSetMapperTest.kt).

I believe this is still inline with the heart/goal of lite-for-jdbc as this does not perform any sql generation magic.

As stated this is experimental as some liberties are being made on Timestamps whereas this mapping tool currently will convert all Timestamp and Time values to java.time.Instant as Timestamp is not an appropriate type to pass around outside of the relational/SQL boundary. There may be an issue with the Calendar on Timestamp, see the extension function `readValue` in _ResultSet.kt